+++ cmd/sks_nodepool_evict.go
@@ -81,7 +81,7 @@ func (c *sksNodepoolEvictCmd) cmdRun(cmd *cobra.Command, _ []string) error {
 		}
 	}
 	if nodepool == nil {
-		return errors.New("Nodepool not found") // nolint:golint
+		return errors.New("Nodepool not found") // nolint:stylecheck
 	}
 
 	nodes := make([]string, len(c.Nodes))